requirements.txt
flask
python-dotenv
flask-cors
gunicorn
pymongo